services/backend/pyproject.toml
[tool.black]
line-length = 120
skip-string-normalization = true
exclude = "/(.git|.tox|migrations|libs|docs|data)/"
target-versions = "py38"


[tool.pdm]
[[tool.pdm.source]]
url = "https://pypi.python.org/simple"
verify_ssl = true
name = "pypi"

[tool.pdm.dev-dependencies]
dev = [
    "black==20.8b1",
    "callee",
    "drf-yasg==1.20.0",
    "factory-boy==2.12.0",
    "flake8",
    "moto[all]==2.0.6",
    "pytest==5.4.1",
    "pytest-django~=3.9",
    "pytest-dotenv",
    "pytest-factoryboy~=2.0",
    "pytest-faker",
    "pytest-freezegun",
    "pytest-mock",
]
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "boto3",
    "Django==3.1.7",
    "djangorestframework~=3.11",
    "djangorestframework-simplejwt==4.6.0",
    "django-environ",
    "django-extensions",
    "django-hashid-field~=3.1",
    "django-hosts",
    "drf-access-policy",
    "dj-stripe",
    "psycopg2-binary",
    "requests",
    "sentry-sdk",
    "six",
    "social-auth-app-django==4.0.0",
    "social-auth-core==4.0.2",
    "whitenoise",
    "graphene-django",
    "gevent~=21.1",
    "graphene-file-upload~=1.3",
    "django-storages~=1.11",
]
requires-python = ">=3.8"
dynamic = ["classifiers"]
license = {text = "MIT"}

[project.urls]
homepage = ""

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

